# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/gamification_badge_users:
    get:
      tags:
        - GamificationBadgeUsers
      summary: List GamificationBadgeUsers
      description: Get a paginated list of GamificationBadgeUsers with optional filtering and sorting
      operationId: listGamificationBadgeUsers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GamificationBadgeUserList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - GamificationBadgeUsers
      summary: Create GamificationBadgeUser
      description: Create a new GamificationBadgeUser
      operationId: createGamificationBadgeUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGamificationBadgeUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GamificationBadgeUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    GamificationBadgeUser:
      type: object
      required:
        - id
        - badge_id
        - recipient_user_id
        - grant_kind
        - granted_at
        - metadata
      properties:
        id:
          type: string
          format: uuid
        badge_id:
          type: string
          format: uuid
        recipient_user_id:
          type: string
          format: uuid
        sender_user_id:
          nullable: true
          type: string
          format: uuid
        grant_kind:
          type: string
        challenge_id:
          nullable: true
          type: string
          format: uuid
        comment:
          nullable: true
          type: string
        level:
          nullable: true
          type: string
        granted_at:
          type: string
          format: date-time
        grant_key:
          nullable: true
          type: string
          maxLength: 200
        metadata:
          type: object
...
    CreateGamificationBadgeUserRequest:
      type: object
      required:
        - badge_id
        - recipient_user_id
        - grant_kind
        - granted_at
        - metadata
      properties:
        badge_id:
          type: string
          format: uuid
        recipient_user_id:
          type: string
          format: uuid
        sender_user_id:
          nullable: true
          type: string
          format: uuid
        grant_kind:
          type: string
        challenge_id:
          nullable: true
          type: string
          format: uuid
        comment:
          nullable: true
          type: string
        level:
          nullable: true
          type: string
        granted_at:
          type: string
          format: date-time
        grant_key:
          nullable: true
          type: string
          maxLength: 200
        metadata:
          type: object
...
    GamificationBadgeUserList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GamificationBadgeUser'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []